Add Square Player
Create a SquarePlayer class and add it to the main game scene. The
regular player has to be removed from the scene because it will cause
interference with the SquarePlayer if left in the scene. This is most
likely a bug with merlin-game-engine.

The serve npm script will change to run webpack before running webpack
serve because webpack serve isn't correctly picking up changes to the
code.