Re-introduce the regular player into the game
Re-add the regular player into the game. The spawnpoint of the regular
player has not been changed, so it still directly overlaps with the
squarePlayer. A game engine update has allowed this to work.

Update to merlin-game-engine v1.0.3. This patch fixes the collision
between two objects that were already inside of eachother. This bug was
the cause of the previous issue with the square player incorrectly
interacting with the regular player.

Increase package.json merlin-game-engine version requirement to
'^1.0.3'. This forces the merlin-game-engine package to update to the
version with the collision fix.