Create basic level change system (#4)
What this will do is transfer the level information from the testGame
file to the newly created level files in the levels folder. Each level
file in the levels folder has a class that returns all of the game
objects which are then called and loaded in the testGame folder.