Level loading system only instantiates one level at a time (#4)
Changes previous basic leveling system by making the system only
instantiate one level at a time. To add more levels, create file in the
levels folder and name file "level" + (# of level). The class inside of
file must be named "Level" + (# of level). In the constructor of
testGame class of the testGame file, add a new object to the array
"levelData" and import the class from the level file you just created.
The level that is used in this version is based off of the variable
"currentLevel" inside of the testGame object. To test a level, you must
change the variable manually.