Create Lever and TogglePlatform
Create a Lever. The lever is a region that whenever a player enters, it
broadcasts the message "togglePlatform", along with which toggle index
to use.

Create a TogglePlatform that reads the "togglePlatform" message and sets
its collider and itself to be toggeled invisible or visible.

Add a lever and togglePlatform to level1 for testing.