Create death conditions for the player
Create the death conditions for the player. The player will die if it is
fully outside of the world, is squished between a staticBody and the
squarePlayer, or touches a spike.

The player detects spikes based on regions that have the word "spike" in
them, ignoring case.

Add a spike to level1 for testing.