Create a Goal object
Create a Goal object that broadcasts the "nextLevel" message. The Goal
takes in a position and size in its constructor. When a player and a
squarePlayer are inside it at the same time, the goal will broadcast the
"nextLevel" message.

Remove the ability to broadcast the "nextLevel" message from the player
for compatability with the new goal system.