Keep the square player's velocity on collision
When the square player collides with the player, its velocity will not
be affected because each frame, it is checking its velocity with its
expected velocity. If the velocities don't match, then the velocity of
the square player will be set to the desired velocity. The desired
veloctity will be reset when there is a collision with a non-player
object along the same axis as the direction of the squareplayer.